+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+ - `TransactionResponse.[get|set]ValidateStatus()`
+ - `TransactionReceipt.validateStatus()`
+ - `TransactionRecord.validateReceiptStatus()`
+ - `TransactionReceipt.transactionId`
+ - `TopicUpdateTransaction.[get|set]ExpirationTime()`
+ - `CustomFee.[set|get]AllCollectorsAreExempt()`
+
 ## 2.17.4
 
 ### Added